refactor instructorAssessments
