+++ pages/instructorAssessments/instructorAssessments.js
@@ -10,14 +10,10 @@ const sanitizeName = require('../../lib/sanitize-name');
 const sqldb = require('@prairielearn/prairielib/sql-db');
 const sqlLoader = require('@prairielearn/prairielib/sql-loader');
 
-const async = require('async');
 const error = require('@prairielearn/prairielib/error');
 const debug = require('debug')('prairielearn:instructorAssessments');
-const fs = require('fs-extra');
-const path = require('path');
-const uuidv4 = require('uuid/v4');
 const logger = require('../../lib/logger');
-const editHelpers = require('../shared/editHelpers');
+const { AssessmentAddEditor } = require('../shared/editHelpers');
 
 const sql = sqlLoader.loadSqlEquiv(__filename);
 
@@ -123,39 +119,22 @@ router.post('/', (req, res, next) => {
     debug(`Responding to post with action ${req.body.__action}`);
     if (req.body.__action == 'add_assessment') {
         debug(`Responding to action add_assessment`);
-
-        if (!res.locals.authz_data.has_course_permission_edit) return next(new Error('Access denied'));
-
-        // Do not allow users to edit the exampleCourse
-        if (res.locals.course.options.isExampleCourse) {
-            return next(error.make(400, `attempting to edit example course`, {
-                locals: res.locals,
-                body: req.body,
-            }));
-        }
-
-        let edit = {
-            userID: res.locals.user.user_id,
-            courseID: res.locals.course.id,
-            coursePath: res.locals.course.path,
-            courseInstanceID: res.locals.course_instance.id,
-            courseInstancePath: path.join(res.locals.course.path, 'courseInstances', res.locals.course_instance.short_name),
-            uid: res.locals.user.uid,
-            user_name: res.locals.user.name,
-        };
-
-        edit.description = 'Add assessment in browser and sync';
-        edit.write = add_write;
-        editHelpers.doEdit(edit, res.locals, (err, job_sequence_id) => {
-            if (ERR(err, (e) => logger.error(e))) {
-                res.redirect(res.locals.urlPrefix + '/edit_error/' + job_sequence_id);
-            } else {
-                debug(`Get assessment_id from tid=${edit.tid} with course_instance_id=${edit.courseInstanceID}`);
-                sqldb.queryOneRow(sql.select_assessment_id_from_tid, {tid: edit.tid, course_instance_id: edit.courseInstanceID}, (err, result) => {
-                    if (ERR(err, next)) return;
-                    res.redirect(res.locals.urlPrefix + '/assessment/' + result.rows[0].assessment_id);
-                });
-            }
+        const editor = new AssessmentAddEditor({
+            locals: res.locals,
+        });
+        editor.canEdit((err) => {
+            if (ERR(err, next)) return;
+            editor.doEdit((err, job_sequence_id) => {
+                if (ERR(err, (e) => logger.error(e))) {
+                    res.redirect(res.locals.urlPrefix + '/edit_error/' + job_sequence_id);
+                } else {
+                    debug(`Get assessment_id from tid=${editor.tid} with course_instance_id=${res.locals.course_instance.id}`);
+                    sqldb.queryOneRow(sql.select_assessment_id_from_tid, {tid: editor.tid, course_instance_id: res.locals.course_instance.id}, (err, result) => {
+                        if (ERR(err, next)) return;
+                        res.redirect(res.locals.urlPrefix + '/assessment/' + result.rows[0].assessment_id);
+                    });
+                }
+            });
         });
     } else {
         next(error.make(400, 'unknown __action: ' + req.body.__action, {
@@ -164,70 +143,5 @@ router.post('/', (req, res, next) => {
         }));
     }
 });
-
-function add_write(edit, callback) {
-    const assessmentsPath = path.join(edit.courseInstancePath, 'assessments');
-    async.series([
-        (callback) => {
-            debug(`Generate unique TID in ${assessmentsPath}`);
-            fs.readdir(assessmentsPath, (err, filenames) => {
-                let number = 1;
-
-                if (err) {
-                    // if the code is ENOENT, then the "assessments" folder does
-                    // not exist, and so there are no assessments yet - otherwise,
-                    // something has gone wrong
-                    if (err.code != 'ENOENT') return ERR(err, callback);
-                } else {
-                    filenames.forEach((filename) => {
-                        let found = filename.match(/^HW([0-9]+)$/);
-                        if (found) {
-                            const foundNumber = parseInt(found[1]);
-                            if (foundNumber >= number) {
-                                number = foundNumber + 1;
-                            }
-                        }
-                    });
-                }
-
-                edit.tid = `HW${number}`;
-                edit.assessmentNumber = number,
-                edit.assessmentPath = path.join(assessmentsPath, edit.tid);
-                edit.pathsToAdd = [
-                    edit.assessmentPath,
-                ];
-                edit.commitMessage = `in-browser edit: add assessment ${edit.tid}`;
-                callback(null);
-            });
-        },
-        (callback) => {
-            debug(`Write infoAssessment.json`);
-
-            // "number" may not be unique - that's ok, the user can change it later -
-            // what's important is that "tid" is unique (see above), because that's a
-            // directory name
-            let infoJson = {
-                uuid: uuidv4(),
-                type: 'Homework',
-                title: 'Replace this title',
-                set: 'Homework',
-                number: `${edit.assessmentNumber}`,
-                allowAccess: [],
-                zones: [],
-            };
-
-            // We use outputJson to create the directory edit.assessmentsPath if it
-            // does not exist (which it shouldn't). We use the file system flag 'wx'
-            // to throw an error if edit.assessmentPath already exists.
-            fs.outputJson(path.join(edit.assessmentPath, 'infoAssessment.json'), infoJson, {spaces: 4, flag: 'wx'}, (err) => {
-                if (ERR(err, callback)) return;
-                callback(null);
-            });
-        },
-    ], (err) => {
-        if (ERR(err, callback)) return;
-        callback(null);
-    });
-}
 
 module.exports = router;
